trama_y_codigo/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
python-dotenv==1.1.0
Werkzeug==3.1.3
